requirements.txt
streamlit>=1.36.0
openai>=1.43.0
chromadb>=0.5.4
pypdf>=4.2.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
langchain-text-splitters>=0.2.2
numpy>=1.26.0


